+++ NEWS.md
@@ -2,6 +2,10 @@
 ## What's changed
  - use `pl$set_polars_options(debug_polars = TRUE)` to profile/debug method-calls of a polars query (#193)
 
+## What's changed
+ - lazy functions translated: `pl$implode`, `pl$explode`, `pl$unique`, `pl$approx_unique`, `pl$head`, `pl$tail` (#196)
+ - `pl$list` is deprecated, use `pl$implode` instead (#196) 
+
 # polars 0.6.0
 ## BREAKING CHANGES
 - Bump rust-polars from 2023-02-17 unreleased version to 2023-04-20 unreleased version. (#183)