+++ NEWS.md
@@ -1,4 +1,4 @@
-# polars (development version)
+# polars 0.6.1
 ## What's changed
  - use `pl$set_polars_options(debug_polars = TRUE)` to profile/debug method-calls of a polars query (#193)